tests/wpt/tests/html/canvas/element/shadows/2d.shadow.attributes.shadowColor.current.removed.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.shadow.attributes.shadowColor.current.removed</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.attributes.shadowColor.current.removed</h1>
<p class="desc">currentColor is solid black when the canvas element is not in a document</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("currentColor is solid black when the canvas element is not in a document");
_addTest(function(canvas, ctx) {

  // Try not to let it undetectably incorrectly pick up opaque-black
  // from other parts of the document:
  document.documentElement.style.color = '#f00';
  document.body.style.color = '#f00';
  canvas.style.color = '#f00';

  canvas.remove();
  ctx.shadowColor = 'currentColor';
  _assertSame(ctx.shadowColor, '#000000', "ctx.shadowColor", "'#000000'");

});
</script>
