LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/shadows/2d.shadow.attributes.shadowColor.current.removed-expected.txt

2d.shadow.attributes.shadowColor.current.removed
currentColor is solid black when the canvas element is not in a document
Actual output:

PASS currentColor is solid black when the canvas element is not in a document
